Fix all the timezone issues (hopefully)
The worksheet.date field is not generated from worksheet.started_at
anymore, but a simple DATE field. This makes much more sense from an
end-user perspective, and is also more in line with the fact that dates
don't have an associated timezone.

The fields defaults to timezone.localdate(), which, with the related
middleware added in a00a019, should always write the correct value. The
only tricky part left is in the tests, but they all pass so everything
looks good (until it all breaks).